refactor(calendar): use DayRange
