+++ calendar/src/main/java/com/jonaswanke/calendar/WeekIndicatorView.kt
@@ -9,6 +9,7 @@ import android.util.AttributeSet
 import android.view.View
 import androidx.annotation.AttrRes
 import androidx.core.content.withStyledAttributes
+import com.jonaswanke.calendar.utils.Day
 import kotlin.properties.Delegates
