+++ example/src/main/java/com/jonaswanke/calendar/example/MainActivity.kt
@@ -12,9 +12,9 @@ import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.GravityCompat
 import androidx.databinding.DataBindingUtil
 import com.jonaswanke.calendar.BaseEvent
-import com.jonaswanke.calendar.Day
+import com.jonaswanke.calendar.utils.Day
 import com.jonaswanke.calendar.Event
-import com.jonaswanke.calendar.Week
+import com.jonaswanke.calendar.utils.Week
 import com.jonaswanke.calendar.example.databinding.ActivityMainBinding
 import kotlinx.android.synthetic.main.activity_main.*
 import java.util.*
